prevent recheck of multiple radio buttons
bug detected when using validation "required" on group of radio buttons with same name attribute, i.e. "payment_method"

when not valid (i.e. no radio button was checked), an error message span would be created for each radio button element.

this patch adds a memory array named checkedInputs[] 
to hold the names of elements that have been checked 
during the iteration of isValid() func,

thereby preventing a recheck of elements with the same name, 
in the situation of radio buttons.

during the iteration, each element is added to this array.

if the element name already exists in the array, the validation of that element will not be performed.

since each form element has unique names, except for radio buttons and checkbox groups, this does not affect other input types.


Important Notes on Multiple Checkboxes:

when using validation on a group of checkboxes, 
set validation rule to checkboxes_group, and specify qty attribute, i.e. min1

otherwise the keyup event will trigger the validation on the checkbox, 
without being able to read the other checkboxes in same named group,
and you will see an error message, even though the required qty of checkboxes is checked