Don't version .dev files; remove code in version task
The mechanism used to version these files is prone to error, because
it does a global search and replace looking
for the prior version and replacing it with the new version.
If any versions have been skipped or any version number has been
 altered by hand, this won't work.
The .dev files are not for release and so do not need to be versioned.
The production files contain a banner which will automatically
pull the right version number when the files are built.