Adding validations for minimum and maximum dates
Changing the date validation function, adding optional attributes min-date and max-date. If informed, it sets an allowed range for the date. It checks following this order: valid date, minimum date, maximum date. Both data-validation-min-date and max-date should be informed using the same format as the field (yyyy-mm-dd as default or any format informed through data-validation-format). English and portuguese locales are ok. The other ones are untranslated.

Request on issue #656